% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLapply.R
\name{chunk}
\alias{chunk}
\title{chunk}
\usage{
chunk(len, threads = getSensibleThreads())
}
\arguments{
\item{len}{the index range to divide}

\item{threads}{the number of chunks to create}
}
\value{
a list (even if there's just one) of lists of (start,end) 
arguments useable via do.call(seq,x)
}
\description{
chunk
}
