% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forkBombGen.R
\name{forkBombGen}
\alias{forkBombGen}
\title{forkBombGen}
\usage{
forkBombGen(FUN, COLLATEFUN = list)
}
\arguments{
\item{FUN}{a single argument function to parallelise.}

\item{COLLATEFUN}{a function used to collate results. Default is list()}
}
\value{
When run with list() as the COLLATEFUN() parameter, a list per
thread run, of lists per argument run through FUN().
}
\description{
A function generator to simplify parallelisation.
}
