% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLapply.R
\name{setChunkSize}
\alias{setChunkSize}
\title{setChunkSize}
\usage{
setChunkSize(x = 2^24)
}
\arguments{
\item{x}{an 'integer' defining a sensible max chunk size.
Default : \code{.Machine$integer.max}}
}
\description{
Set the served chunkSize
}
