% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLapply.R
\name{setSensibleThreads}
\alias{setSensibleThreads}
\title{setSensibleThreads}
\usage{
setSensibleThreads(x = max(1, parallel::detectCores() - 1))
}
\arguments{
\item{x}{a positive 'integer' defining a sensible number of threads to run. 
Default : \code{max(1,parallel::detectCores()-1)}}
}
\description{
Get the sensible number of threads to plan to run concurrently.
}
