% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{chunk
Trying to set both threads and chunkSize will return an error. 
If you know both, you don't need this.}
\usage{
chunk(from = 1, to, chunkSize = NULL, threads = NULL)
}
\arguments{
\item{from}{the start of the index range to chunk}

\item{to}{the end of the index range to chunk}

\item{chunkSize}{the size of chunks to create}

\item{threads}{the number of chunks to create}
}
\value{
a list (even if there's just one) of lists of (from,to)  
arguments useable via do.call(seq,x)
}
\description{
chunk
Trying to set both threads and chunkSize will return an error. 
If you know both, you don't need this.
}
