% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLapply.R
\name{chunk}
\alias{chunk}
\title{chunk}
\usage{
chunk(len, threads = getSensibleThreads())
}
\arguments{
\item{len}{the index range to divide}

\item{threads}{the number of chunks to create}
}
\value{
a list of c(start,end) vectors defining chunks
}
\description{
chunk
}
